refactoring to divide omezarr metadata and utility functions
